added

+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: aeminian <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/19 22:06:01 by aeminian          #+#    #+#             */
-/*   Updated: 2024/01/19 22:06:18 by aeminian         ###   ########.fr       */
+/*   Updated: 2024/01/20 00:56:45 by adel             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ char	*ft_strncat(char *src, char *dest, unsigned int n)
 		i++;
 	}
 	j = 0;
-	while (dest[j] && n > j )
+	while (dest[j] && (int)n > j )
 	{
 		src[i] = dest[j];
 		i++;
